added trace and untrace commands

+++ esp32/tiny/objExpression.cpp
@@ -210,7 +210,7 @@ calcCount = 0;
           break;
 		  
 		  case '^':		  
-		  z = pow(x, y);		  
+		  z = float(pow(x, y));		  
 		  break;
       }
